Controlling Access with Token Abilities.

+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -9,10 +9,14 @@ use App\Http\Requests\Api\V1\UpdateTicketRequest;
 use App\Http\Resources\Api\TicketResource;
 use App\Models\Ticket;
 use App\Models\User;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Support\Facades\Gate;
 
 class TicketController extends ApiController
 {
+    protected string $policyNamespace = 'App\\Policies\\V1';
+
     /**
      * Display a listing of the resource.
      */
@@ -28,6 +32,9 @@ class TicketController extends ApiController
     {
         try {
             $user = User::findOrFail($request->input('data.relationships.author.data.id'));
+
+            Gate::authorize('store', null);
+
         } catch (ModelNotFoundException $exception) {
             return $this->ok(
             'User not found.',
@@ -69,12 +76,16 @@ class TicketController extends ApiController
         try {
             $ticket = Ticket::findOrFail($ticket_id);
 
+            Gate::authorize('replace', [$ticket]);
+
             $ticket->update($request->mappedAttributes());
 
             return TicketResource::make($ticket);
 
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
+        } catch (AuthorizationException $e) {
+            return $this->error(  "You are not authorized to replace that resource", 401);
         }
     }
 
@@ -86,12 +97,16 @@ class TicketController extends ApiController
         try {
             $ticket = Ticket::findOrFail($ticket_id);
 
+            Gate::authorize('update', [$ticket]);
+
             $ticket->update($request->mappedAttributes());
 
             return TicketResource::make($ticket);
 
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
+        } catch (AuthorizationException $e) {
+            return $this->error(  "You are not authorized to update that resource", 401);
         }
     }
 
@@ -102,11 +117,16 @@ class TicketController extends ApiController
     {
         try {
             $ticket = Ticket::findOrFail($ticket_id);
+
+            Gate::authorize('delete', [$ticket]);
+
             $ticket->delete();
 
             return $this->ok('Ticket successfully deleted.');
         } catch (ModelNotFoundException $exception) {
             return $this->error('Ticket cannot be found.', 404);
+        } catch (AuthorizationException $e) {
+            return $this->error(  "You are not authorized to delete that resource", 401);
         }
     }
 }
